refactor: extract Retryable into a Feature interface
Move retry knobs off the base Adapter interface — they only make sense
for adapters that actually own a network connection. Filesystem, Memory,
None, Pool, Sharding, and CircuitBreaker no longer carry no-op or
fan-out retry methods. Retryable is now a Feature interface implemented
by Redis, RedisCluster, Memcached, and Hazelcast.

Multiplexing drops its configurable retry knobs (configurable retry
across a shared connection drops all in-flight callers anyway) but
keeps a single transparent reconnect-and-retry on ConnectionException.

Also fix Multiplexing::save() swallowing connection exceptions — encode
errors now surface the underlying JsonException instead of being masked
as a false return.
