+++ src/Cache/Adapter/Memory.php
@@ -19,24 +19,6 @@ class Memory implements Adapter
     }
 
     /**
-     * @param  int  $maxRetries (0-10)
-     * @return self
-     */
-    public function setMaxRetries(int $maxRetries): self
-    {
-        return $this;
-    }
-
-    /**
-     * @param  int  $retryDelay time in milliseconds
-     * @return self
-     */
-    public function setRetryDelay(int $retryDelay): self
-    {
-        return $this;
-    }
-
-    /**
      * @param  string  $key
      * @param  int  $ttl
      * @param  string  $hash optional
@@ -155,21 +137,5 @@ class Memory implements Adapter
     public function getName(?string $key = null): string
     {
         return 'memory';
-    }
-
-    /**
-     * @return int
-     */
-    public function getMaxRetries(): int
-    {
-        return 0;
-    }
-
-    /**
-     * @return int
-     */
-    public function getRetryDelay(): int
-    {
-        return 0;
     }
 }
