+++ src/Cache/Adapter/None.php
@@ -14,24 +14,6 @@ class None implements Adapter
     }
 
     /**
-     * @param  int  $maxRetries (0-10)
-     * @return self
-     */
-    public function setMaxRetries(int $maxRetries): self
-    {
-        return $this;
-    }
-
-    /**
-     * @param  int  $retryDelay time in milliseconds
-     * @return self
-     */
-    public function setRetryDelay(int $retryDelay): self
-    {
-        return $this;
-    }
-
-    /**
      * @param  string  $key
      * @param  int  $ttl
      * @param  string  $hash optional
@@ -113,21 +95,5 @@ class None implements Adapter
     public function getName(?string $key = null): string
     {
         return 'none';
-    }
-
-    /**
-     * @return int
-     */
-    public function getMaxRetries(): int
-    {
-        return 0;
-    }
-
-    /**
-     * @return int
-     */
-    public function getRetryDelay(): int
-    {
-        return 0;
     }
 }
